+++ 3483/3483_2/bot_start.rb
@@ -1,0 +1,18 @@
+require 'telegram/bot'
+
+Telegram::Bot::UpdatesController.session_store = :redis_store, { expires_in: 1.month }
+
+class WebhooksController < Telegram::Bot::UpdatesController
+  include Telegram::Bot::UpdatesController::MessageContext
+  include Telegram::Bot::UpdatesController::Session
+end
+
+TOKEN = ENV['BOT_TOKEN']
+bot = Telegram::Bot::Client.new(TOKEN)
+
+
+# poller-mode
+require 'logger'
+logger = Logger.new(STDOUT)
+poller = Telegram::Bot::UpdatesPoller.new(bot, WebhooksController, logger: logger)
+poller.start
