--- 3483/3483_2/config/locales/en.yml
@@ -1,14 +1,0 @@
-en:
-  already_registered: 'You are already registered.'
-  give_number: 'Hello! Give me your personal number.'
-  succesfull_registration: 'Success! Registration completed.'
-  wrong_number: 'Wrong number. Try again.'
-  selfie: 'Send me selfie, please.'
-  not_selfie: "Sorry, it doesn't look like selfie. Try again."
-  not_registered: 'You are not registered. Type /start to start registration.'
-  already_checkin: 'You are already checkin.'
-  not_checkin: 'You are not checkin. Type /checkin to start checkin.'
-  geo: 'Send me your geoposition.'
-  not_geo: 'It seems, you are not in a camp. Try again.'
-  success: 'Success!'
-
