--- 3483/3483_2/data/numbers.txt
@@ -1,72 +1,0 @@
-3475
-3754
-3581
-3548
-3481
-3745
-3493
-3469
-3741
-3634
-3506
-3710
-3501
-3723
-3582
-3737
-3478
-3495
-3515
-3498
-3614
-3542
-3496
-3485
-3600
-3730
-3725
-3483
-3604
-3648
-3684
-3587
-3761
-3530
-3561
-3750
-3696
-3567
-3467
-3536
-3532
-3460
-3459
-3573
-3529
-3488
-3486
-3547
-3476
-3531
-3523
-3457
-3626
-3473
-3574
-3688
-3638
-3541
-3570
-3657
-3512
-3444
-3715
-3711
-3640
-3729
-3123
-3496
-3538
-3520
-3545
-3663
