--- 3483/3483_2/validations/helper.rb
@@ -1,11 +1,0 @@
-module Validations
-  module Helper
-    def checkin?
-      chat_session[session_key]['checkin']
-    end
-
-    def not_checkin?
-      !checkin?
-    end
-  end
-end
