--- 3483/3483_2/validations/registration.rb
@@ -1,7 +1,0 @@
-module Validations
-  module Registration
-    def registered?
-      chat_session[session_key]
-    end
-  end
-end
