--- 3483/3483_2/validations/selfie.rb
@@ -1,7 +1,0 @@
-module Validations
-  module Selfie
-    def selfie?
-      payload['photo']
-    end
-  end
-end
