--- 3483/3483_2/validations/validations.rb
@@ -1,7 +1,0 @@
-require_relative 'registration.rb'
-require_relative 'selfie.rb'
-require_relative 'geo.rb'
-require_relative 'helper.rb'
-
-module Validations
-end
